pyproject.toml
[project]
name = "msigma-django-assignment-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "faker>=40.5.1",
    "flower>=2.0.1",
    "gunicorn>=25.1.0",
    "psycopg2-binary>=2.9.11",
    "redis>=7.2.1",
    "requests>=2.32.5",
    "whitenoise>=6.12.0",
]
